+++ pex/pex_info.py
@@ -11,11 +11,29 @@ from collections import namedtuple
 
 from .common import open_zip
 from .compatibility import string as compatibility_string
+from .compatibility import PY2
 from .orderedset import OrderedSet
 
 PexPlatform = namedtuple('PexPlatform', 'interpreter version strict')
 
 
+def process_bool(value, negate=False):
+  def apply_bool(pex_info, env_variable):
+    if env_variable.strip().lower() in ('0', 'false'):
+      pex_info[value] = True if negate else False
+    elif env_variable.strip().lower() in ('1', 'true'):
+      pex_info[value] = False if negate else True
+    else:
+      raise ValueError('Unknown value for %s: %r' % (env_variable, value))
+
+
+def process_string(value):
+  def apply_string(pex_info, env_variable):
+    pex_info[value] = env_variable
+  return apply_string
+
+
+# TODO(wickman) Split this into a PexInfoBuilder/PexInfo to ensure immutability.
 class PexInfo(object):
   """PEX metadata.
 
@@ -24,18 +42,32 @@ class PexInfo(object):
   code_hash: str                    # sha1 hash of all names/code in the archive
   distributions: {dist_name: str}   # map from distribution name (i.e. path in
                                     # the internal cache) to its cache key (sha1)
+  requirements: list                # list of requirements for this environment
 
   # Environment options
   pex_root: ~/.pex                   # root of all pex-related files
+                                     # PEX_ROOT
+
   entry_point: string                # entry point into this pex
+                                     # PEX_MODULE
+
+  script: string                     # script to execute in this pex environment
+                                     # at most one of script/entry_point can be specified
+                                     # PEX_SCRIPT
+
   zip_safe: True, default False      # is this pex zip safe?
+                                     # PEX_FORCE_LOCAL
+
   inherit_path: True, default False  # should this pex inherit site-packages + PYTHONPATH?
+                                     # PEX_INHERIT_PATH
+
   ignore_errors: True, default False # should we ignore inability to resolve dependencies?
+                                     # PEX_IGNORE_ERRORS
+
   always_write_cache: False          # should we always write the internal cache to disk first?
                                      # this is useful if you have very large dependencies that
                                      # do not fit in RAM constrained environments
-  requirements: list                 # list of requirements for this environment
-
+                                     # PEX_ALWAYS_CACHE
 
   .. versionchanged:: 0.8
     Removed the ``repositories`` and ``indices`` information, as they were never
@@ -44,6 +76,15 @@ class PexInfo(object):
 
   PATH = 'PEX-INFO'
   INTERNAL_CACHE = '.deps'
+  ENVIRONMENT_VARIABLES = {
+      'PEX_ROOT': process_string('pex_root'),
+      'PEX_MODULE': process_string('entry_point'),
+      'PEX_SCRIPT': process_string('script'),
+      'PEX_FORCE_LOCAL': process_bool('zip_safe', negate=True),
+      'PEX_INHERIT_PATH': process_string('inherit_path'),
+      'PEX_IGNORE_ERRORS': process_bool('ignore_errors'),
+      'PEX_ALWAYS_CACHE': process_bool('always_write_cache'),
+  }
 
   @classmethod
   def make_build_properties(cls):
@@ -62,7 +103,6 @@ class PexInfo(object):
     pex_info = {
       'requirements': [],
       'distributions': {},
-      'always_write_cache': False,
       'build_properties': cls.make_build_properties(),
     }
     return cls(info=pex_info)
@@ -81,7 +121,16 @@ class PexInfo(object):
   def from_json(cls, content):
     if isinstance(content, bytes):
       content = content.decode('utf-8')
-    return PexInfo(info=json.loads(content))
+    return cls(info=json.loads(content))
+
+  @classmethod
+  def from_env(cls):
+    pex_info = {}
+    for variable, processor in cls.ENVIRONMENT_VARIABLES.items():
+      if variable in os.environ:
+        cls.debug('processing %s = %s' % (variable, os.environ[variable]))
+        processor(pex_info, os.environ[variable])
+    return cls(info=pex_info)
 
   @classmethod
   def _parse_requirement_tuple(cls, requirement_tuple):
@@ -113,6 +162,12 @@ class PexInfo(object):
       raise ValueError('Expected requirements to be a list, got %s' % type(requirements))
     self._requirements = OrderedSet(self._parse_requirement_tuple(req) for req in requirements)
 
+  def _get_safe(self, key):
+    if key not in self._pex_info:
+      return None
+    value = self._pex_info[key]
+    return value.encode('utf-8') if PY2 else value
+
   @property
   def build_properties(self):
     """Information about the system on which this PEX was generated.
@@ -139,9 +194,6 @@ class PexInfo(object):
     By default zip_safe is True.  May be overridden at runtime by the $PEX_FORCE_LOCAL environment
     variable.
     """
-    if 'PEX_FORCE_LOCAL' in os.environ:
-      self.debug('PEX_FORCE_LOCAL forcing zip_safe to False')
-      return False
     return self._pex_info.get('zip_safe', True)
 
   @zip_safe.setter
@@ -158,11 +210,7 @@ class PexInfo(object):
     By default inherit_path is False.  This may be overridden at runtime by the $PEX_INHERIT_PATH
     environment variable.
     """
-    if 'PEX_INHERIT_PATH' in os.environ:
-      self.debug('PEX_INHERIT_PATH override detected')
-      return True
-    else:
-      return self._pex_info.get('inherit_path', False)
+    return self._pex_info.get('inherit_path', False)
 
   @inherit_path.setter
   def inherit_path(self, value):
@@ -186,14 +234,19 @@ class PexInfo(object):
 
   @property
   def entry_point(self):
-    if 'PEX_MODULE' in os.environ:
-      self.debug('PEX_MODULE override detected: %s' % os.environ['PEX_MODULE'])
-      return os.environ['PEX_MODULE']
-    return self._pex_info.get('entry_point')
+    return self._get_safe('entry_point')
 
   @entry_point.setter
   def entry_point(self, value):
     self._pex_info['entry_point'] = value
+
+  @property
+  def script(self):
+    return self._get_safe('script')
+
+  @script.setter
+  def script(self, value):
+    self._pex_info['script'] = value
 
   def add_requirement(self, requirement):
     self._requirements.add(str(requirement))
@@ -211,9 +264,6 @@ class PexInfo(object):
 
   @property
   def always_write_cache(self):
-    if 'PEX_ALWAYS_CACHE' in os.environ:
-      self.debug('PEX_ALWAYS_CACHE override detected: %s' % os.environ['PEX_ALWAYS_CACHE'])
-      return True
     return self._pex_info.get('always_write_cache', False)
 
   @always_write_cache.setter
@@ -240,6 +290,13 @@ class PexInfo(object):
   @property
   def zip_unsafe_cache(self):
     return os.path.join(self.pex_root, 'code')
+
+  def update(self, other):
+    if not isinstance(other, PexInfo):
+      raise TypeError('Cannot merge a %r with PexInfo' % type(other))
+    self._pex_info.update(other._pex_info)
+    self._distributions.update(other.distributions)
+    self._requirements.update(other.requirements)
 
   def dump(self):
     pex_info_copy = self._pex_info.copy()